countmaskr: Small Cell Masking Tool for One- & Two-Way Tabular Reports

Provides automated small-cell suppression for one- and two-way frequency tables. Cells falling below a user-defined frequency threshold are masked, with suppression propagated to secondary cells to prevent indirect disclosure. Designed for clinical and health administrative data, the package supports a range of tabular structures and fits into reproducible reporting pipelines, reducing manual review while applying consistent suppression rules across data sharing workflows.

See also

Author

Maintainer: Sahal Master sahalmaster@outlook.com (ORCID)

Authors:

Amy Goodwin Davies goodwindaa@chop.edu

Allison Zelinski zelinskia@chop.edu

Qiwei Shen shenq@chop.edu (ORCID)

Other contributors:

Nicole Marchesani MARCHESANN@chop.edu (ORCID) [contributor]

Aqsa Khan khana8@chop.edu (ORCID) [contributor]

Rhonda DeCook rhonda-decook@uiowa.edu [contributor]